Support images in nested forms
Previously, images in nested forms wouldn't be able to be cropped, as
the `croppables` param built by the form helper wasn't compatible with
Rails' nested form structure.

This changes the structure produced by the form helper to leverage
Rails' nested forms feature. In doing so, we also remove
Croppable::Param and CleanCroppableParams in favor of passing the
(potentially nested) attributes through to the attribute setter as a
hash.

Note that to use Croppable with strong parameters,
Croppable::PermittedParams should be included in the controller. This
intentionally does not include it by default when the engine mounts.