Webpack[er] support
Previously, ESM modules / sprockets was the only way to include the
files in a Rails project.

The official way to include gem-based javascript in a webpacker-based
Rails project is to create an npm module, see ActionCable as an example:
https://github.com/rails/rails/tree/main/actioncable

This adds webpacker support by copying the ActionCable approach,
including using rollup to transpile a plain JS file into ESM and UMD
versions.

Related, this also makes the croppable javascript export a function
to invoke the croppable functionality instead of hardcoding callbacks on
DOMContentLoaded and turbo:load events, which lets the project be more
flexible in how it's instantiated (ex. via turbolinks instead of turbo).