Don't automatically include croppable in models
Maybe it's a personal preference, but I don't love it when I inspect a
model and see tons of unrelated functionality included by plugins. I
prefer to include behavior where it's needed, which is also nice when
looking at a model to see what it does. The end developer can always
include the module globally in an initializer, of course.