Support inline job runners
When performing an inline job, such as in a test or by choice in an app,
we only have the uploaded file; the original will not be written to
the storage service until after_commit, and we can't rely on croppable's
after commit coming after Rails' after_commit (in practice it sometimes
does, sometimes doesn't, depending on the attachment). Conversely, after
the file has been saved to storage and we're in the background job, the
tempfile has been cleaned up and we only have the stored file. So, we
look for the tempfile, and if not found use the stored file.